@@ -10,11 +10,6 @@ import 'dart:collection';
 /// If [start] `==` [target], an empty [List] is returned and [edges] is never
 /// called.
 ///
-/// [start], [target] and all values returned by [edges] must not be `null`.
-/// If asserts are enabled, an [AssertionError] is raised if these conditions
-/// are not met. If asserts are not enabled, violations result in undefined
-/// behavior.
-///
 /// If [equals] is provided, it is used to compare nodes in the graph. If
 /// [equals] is omitted, the node's own [Object.==] is used instead.
 ///
@@ -24,12 +19,12 @@ import 'dart:collection';
 ///
 /// If you supply one of [equals] or [hashCode], you should generally also to
 /// supply the other.
-Iterable<T> shortestPath<T>(
+Iterable<T>? shortestPath<T extends Object>(
   T start,
   T target,
   Iterable<T> Function(T) edges, {
-  bool Function(T, T) equals,
-  int Function(T) hashCode,
+  bool Function(T, T)? equals,
+  int Function(T)? hashCode,
 }) =>
     _shortestPaths<T>(
       start,
@@ -58,11 +53,11 @@ Iterable<T> shortestPath<T>(
 ///
 /// If you supply one of [equals] or [hashCode], you should generally also to
 /// supply the other.
-Map<T, Iterable<T>> shortestPaths<T>(
+Map<T, Iterable<T>> shortestPaths<T extends Object>(
   T start,
   Iterable<T> Function(T) edges, {
-  bool Function(T, T) equals,
-  int Function(T) hashCode,
+  bool Function(T, T)? equals,
+  int Function(T)? hashCode,
 }) =>
     _shortestPaths<T>(
       start,
@@ -71,21 +66,20 @@ Map<T, Iterable<T>> shortestPaths<T>(
       hashCode: hashCode,
     );
 
-Map<T, Iterable<T>> _shortestPaths<T>(
+Map<T, Iterable<T>> _shortestPaths<T extends Object>(
   T start,
   Iterable<T> Function(T) edges, {
-  T target,
-  bool Function(T, T) equals,
-  int Function(T) hashCode,
+  T? target,
+  bool Function(T, T)? equals,
+  int Function(T)? hashCode,
 }) {
-  assert(start != null, '`start` cannot be null');
-  assert(edges != null, '`edges` cannot be null');
-
   final distances = HashMap<T, _Tail<T>>(equals: equals, hashCode: hashCode);
   distances[start] = _Tail<T>();
 
-  equals ??= _defaultEquals;
-  if (equals(start, target)) {
+  final nonNullEquals = equals ??= _defaultEquals;
+  final isTarget =
+      target == null ? _neverTarget : (T node) => nonNullEquals(node, target);
+  if (isTarget(start)) {
     return distances;
   }
 
@@ -93,15 +87,14 @@ Map<T, Iterable<T>> _shortestPaths<T>(
 
   while (toVisit.isNotEmpty) {
     final current = toVisit.removeFirst();
-    final currentPath = distances[current];
+    final currentPath = distances[current]!;
 
     for (var edge in edges(current)) {
-      assert(edge != null, '`edges` cannot return null values.');
       final existingPath = distances[edge];
 
       if (existingPath == null) {
         distances[edge] = currentPath.append(edge);
-        if (equals(edge, target)) {
+        if (isTarget(edge)) {
           return distances;
         }
         toVisit.add(edge);
@@ -113,6 +106,7 @@ Map<T, Iterable<T>> _shortestPaths<T>(
 }
 
 bool _defaultEquals(Object a, Object b) => a == b;
+bool _neverTarget(Object _) => false;
 
 /// An immutable iterable that can efficiently return a copy with a value
 /// appended.
@@ -123,9 +117,9 @@ bool _defaultEquals(Object a, Object b) => a == b;
 /// space because it copies all the values to a new list and uses that
 /// iterator in order to avoid stack overflows for large paths. This copy is
 /// cached for subsequent calls.
-class _Tail<T> extends Iterable<T> {
-  final T /*?*/ tail;
-  final _Tail<T> /*?*/ head;
+class _Tail<T extends Object> extends Iterable<T> {
+  final T? tail;
+  final _Tail<T>? head;
   @override
   final int length;
   _Tail()
@@ -135,19 +129,16 @@ class _Tail<T> extends Iterable<T> {
   _Tail._(this.tail, this.head, this.length);
   _Tail<T> append(T value) => _Tail._(value, this, length + 1);
 
-  Iterator<T> /*?*/ _iterator;
-
   @override
-  Iterator<T> get iterator {
-    if (_iterator == null) {
-      var /*_Tail<T>?*/ next = this;
-      var values = List<T>.generate(length, (_) {
-        var val = next.tail;
-        next = next.head;
-        return val;
-      });
-      _iterator = values.reversed.iterator;
-    }
-    return _iterator;
-  }
+  Iterator<T> get iterator => _asIterable.iterator;
+
+  late final _asIterable = () {
+    _Tail<T>? next = this;
+    var reversed = List.generate(length, (_) {
+      var val = next!.tail;
+      next = next!.head;
+      return val as T;
+    });
+    return reversed.reversed;
+  }();
 }
