@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-bool xEquals(X a, X b) => a?.value == b?.value;
+bool xEquals(X a, X b) => a.value == b.value;
 
 int xHashCode(X a) => a.value.hashCode;
 
